Extract reusable scenarios out of integTest
These scenarios will later be used to drive unit tests of the
same functionality.
